UI update

+++ car-rental-common/Classes/Booking.cs
@@ -5,15 +5,15 @@ namespace car_rental_common.Classes;
 
 public class Booking : IBooking
 {
-    public int? Odometer { get; set; }
-    public float? KmDriven { get; set; }
+    public decimal? Odometer { get; set; }
+    public decimal? KmDriven { get; set; }
     public DateTime RentDate { get; set; }
     public DateTime ReturnDate { get; set; }
     public IVehicle Vehicle { get; set; }
     public ICustomer Customer { get; set; }
     public VehicleStatuses Status { get; set; }
 
-    public Booking(int? odometer, float? kmDriven, IVehicle vehicle, ICustomer customer, DateTime rentDate, DateTime returnDate)
+    public Booking(decimal? odometer, decimal? kmDriven, IVehicle vehicle, ICustomer customer, DateTime rentDate, DateTime returnDate)
     {
         Odometer = odometer;
         KmDriven = kmDriven;
