+++ car-rental-common/Classes/Vechicle.cs
@@ -7,9 +7,9 @@ public class Vehicle : IVehicle
 {
     public string RegNo { get; set; }
     public string Make { get; set; }
-    public int? Odometer { get; set; }
-    public float? CostKm { get; set; }
-    public int CostDay { get; set; }
+    public decimal? Odometer { get; set; }
+    public decimal? CostKm { get; set; }
+    public decimal CostDay { get; set; }
     public VehicleStatuses Status { get; set; }
     private VehicleTypes _vehicleType;
     public VehicleTypes VehicleType
@@ -22,7 +22,7 @@ public class Vehicle : IVehicle
         }
     }
 
-    public Vehicle(string regNo = "", string make = "", int? odometer = null, float? costKm = null, VehicleTypes vehicleType = VehicleTypes.Other, VehicleStatuses status = VehicleStatuses.Available)
+    public Vehicle(string regNo = "", string make = "", decimal? odometer = null, decimal? costKm = null, VehicleTypes vehicleType = VehicleTypes.Other, VehicleStatuses status = VehicleStatuses.Available)
     {
         RegNo = regNo;
         Make = make;
