+++ car-rental-common/Extensions/ExtensionMethods.cs
@@ -2,7 +2,7 @@
 
 public static class ExtensionMethods
 {
-    public static int Duration(this DateTime rentDate, DateTime returnDate)
+    public static decimal Duration(this DateTime rentDate, DateTime returnDate)
     {
         TimeSpan duration = returnDate - rentDate;
         return duration.Days + 1;
