+++ car-rental-common/Interfaces/IBooking.cs
@@ -4,8 +4,8 @@ namespace car_rental_common.Interfaces;
 
 public interface IBooking
 {
-    int? Odometer { get; set; }
-    float? KmDriven { get; set; }
+    decimal? Odometer { get; set; }
+    decimal? KmDriven { get; set; }
     DateTime RentDate { get; set; }
     DateTime ReturnDate { get; set; }
     IVehicle Vehicle { get; set; }
