+++ car-rental-common/Interfaces/IVehicle.cs
@@ -6,9 +6,9 @@ public interface IVehicle
 {
     string RegNo { get; set; }
     string Make { get; set; }
-    int? Odometer { get; set; }
-    float? CostKm { get; set; }
-    int CostDay { get; set; }
+    decimal? Odometer { get; set; }
+    decimal? CostKm { get; set; }
+    decimal CostDay { get; set; }
     VehicleTypes VehicleType { get; set; }
     VehicleStatuses Status { get; set; }
 }
